Finishing and testing destroy method in housecontroller (validation refactoring)

+++ src/controllers/HouseController.js
@@ -26,6 +26,7 @@ class HouseController {
         const { user_id } = req.headers
 
 
+
         const house = await House.create({
             user: user_id,
             thumbnail: filename,
@@ -47,12 +48,12 @@ class HouseController {
         const { description, price, location, status } = req.body
         const { user_id } = req.headers
 
-       const user = await User.findById(user_id)
-       const house = await House.findById(house_id)
+        const user = await User.findById(user_id)
+        const house = await House.findById(house_id)
 
-       if(String(user._id) != String(house.user)) {
-           return res.status(401).json( {Error: 'Not authorized'})
-       }
+        if (String(user._id) != String(house.user)) {
+            return res.status(401).json({ Error: 'Not authorized' })
+        }
 
         const housesUpdate = await House.updateOne({ _id: house_id }, {
             user: user_id,
@@ -63,13 +64,22 @@ class HouseController {
             status,
         })
 
-        return res.json(housesUpdate)
+        return res.send()
     }
 
     //Delete house
     async destroy(req, res) {
         const { house_id } = req.params
-        const DeleteHouse = await House.findByIdAndDelete({ _id: house_id })
+        const { user_id } = req.headers
+
+        const getHouse = await House.findById(house_id)
+        const getUser = await User.findById(user_id)
+
+        if (String(getHouse.user) != String(getUser._id)) {
+            return res.status(404).json({ Error: 'Not authorized' })
+        }
+        let DeleteHouse
+        DeleteHouse = await House.findByIdAndDelete({ _id: house_id })
         return res.json(DeleteHouse)
     }
 }
